% Generated by roxygen2 (4.0.2): do not edit by hand
\name{c_project_extractor}
\alias{c_project_extractor}
\title{project_extractor}
\usage{
c_project_extractor(urls)
}
\arguments{
\item{urls}{a vector of URLs}
}
\value{
a vector of language and project names, or "Unknown" if a URL cannot be parsed. In the event
that a URL can be parsed but you've been silly enough to pass it a non-Wikimedia URL, it will simply
return nonsense.
}
\description{
extracts language and project from a Wikimedia URL
}
\details{
\code{project_extractor} takes Wikimedia URLs and extracts the language and project
(for example, turning "https://en.wikipedia.org"" into "en.wikipedia"). It can handle both current
and historic intermediary domains - zero, mobile, wap - and exclude them consistently.
}
\seealso{
\code{\link{host_handler}} for extracting hostnames generically.
}

