% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cryptohash}
\alias{cryptohash}
\title{cryptohash}
\usage{
cryptohash(x, algorithm, include_rand = FALSE, na_rm = FALSE)
}
\arguments{
\item{x}{a vector of strings, or an object that can be coerced into a vector of strings.}

\item{algorithm}{the hashing algorithm to apply to each string. Options are "md5", "sha1", "sha256",
and "sha512".}

\item{include_rand}{whether or not to include a pseudorandom element as part of each string; set to FALSE
by default.}

\item{na_rm}{whether or not it is acceptable to remove NA values from the vector. Please note that
in the event that NA values are found, and na.rm is FALSE, the function call will simply terminate.
Set to FALSE by default.}
}
\value{
a vector of cryptographic hashes, the length of the input vector (or shorter if
na_rm is set to TRUE)
}
\description{
cryptographically hash a character vector, or something coercable into one.
}
\examples{
cryptohash("foo", algorithm = "md5")
cryptohash("foo", algorithm = "md5", include_rand = TRUE)
cryptohash(c("foo",NA), algorithm = "sha1", na_rm = TRUE)
}

