% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dep_hive_query}
\alias{dep_hive_query}
\title{dep_hive_query}
\usage{
dep_hive_query(query, file = NULL, dt = TRUE, ...)
}
\arguments{
\item{query}{a query, or the location of a .hql file containing a query.}

\item{file}{a file name. If this is provided, the results of the query will be written straight
there, and a boolean TRUE returned. If not provided (it's NULL by default), the results of the query
will be returned as a data.frame}

\item{dt}{Whether to return it as a data.table or not.}

\item{...}{other arguments to pass to read.delim.}
}
\value{
a data.frame containing the results of the query, or a boolean TRUE if the user has chosen
to write straight to file.
}
\description{
this is the "old" hive querying function - it's deprecated as all hell and waiting
until Andrew sticks the hive server on a dedicated and more powerful machine.
}
\details{
the deprecated hive querying function
}
\section{escaping}{

\code{hive_query} works by running the query you provide through the CLI via a system() call.
As a result, single escapes for meaningful characters (such as quotes) within the query will not work:
R will interpret them only as escaping that character /within R/. Double escaping (\\\) is thus necessary,
in the same way that it is for regular expressions.
}

\section{handling our hadoop/hive setup}{


The \code{webrequests} table is documented
\href{https://wikitech.wikimedia.org/wiki/Analytics/Cluster/Hive}{on Wikitech}, which also provides
\href{https://wikitech.wikimedia.org/wiki/Analytics/Cluster/Hive/Queries}{a set of example
queries}.

When it comes to manipulating the rows with Java before they get to you, Nuria has written a
\href{https://wikitech.wikimedia.org/wiki/Analytics/Cluster/Hive/QueryUsingUDF}{brief tutorial on loading UDFs}
which should help if you want to engage in that; the example provided is a user agent parser, allowing you to
get the equivalent of \code{\link{ua_parse}}'s output further upstream.
}
\seealso{
\code{\link{log_strptime}} for converting the "dt" column in the webrequests table to POSIXlt,
and \code{\link{mysql_query}} and \code{\link{global_query}} for querying our MySQL databases.
}

