% Generated by roxygen2 (4.0.2): do not edit by hand
\name{device_classifier}
\alias{device_classifier}
\title{device_classifier}
\usage{
device_classifier(devices)
}
\arguments{
\item{devices}{a vector of devices, output by ua_parse}
}
\value{
a character vector of "Phone", "Tablet" or "Other"
}
\description{
classifies devices that come out of ua_parse
}
\seealso{
\code{\link{ua_parse}} for user agent parsing.
}

