% Generated by roxygen2 (4.0.2): do not edit by hand
\name{devices}
\alias{devices}
\title{Device classifiers.}
\format{two character vectors:
\describe{
  \item{mobile_devices}{devices that are phones.}
  \item{tablet_devices}{devices that are tablets}
}}
\description{
A dataset that maps \code{\link{ua_parse}}'s "device" column to device classes - "tablet","phone" or "other".
}

