% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_mcc}
\alias{extract_mcc}
\title{extract_mcc}
\usage{
extract_mcc(x_analytics)
}
\arguments{
\item{x_analytics}{a vector of x_analytics fields}
}
\value{
a vector of empty strings or MCC codes, depending on if an MCC was found.
}
\description{
extract a mobile carrier code (mcc) from an x_analytics header
}
\details{
\code{\link{extract_mcc}} grabs a mobile carrier code and associated country code
from an x_analytics field, where such a code is present (i.e., requests from Zero)
}

