% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geo_city}
\alias{geo_city}
\title{geo_city}
\usage{
geo_city(ip_addresses)
}
\arguments{
\item{ips}{a vector of IP addresses. These will be processed through \code{xff_handler}
before being run, so don't worry if they're a bit groaty.}
}
\value{
a vector of city names. NULL or invalid responses from the API will be replaced with the string "Invalid".
}
\description{
\code{geo_city} geolocates IPv4 and IPv6 addresses to provide city-level results, providing
the English-language name for that city. It uses
\href{http://dev.maxmind.com/geoip/}{MaxMind's binary geolocation database} - the only
limitation is that accuracy
\href{http://www.maxmind.com/en/city_accuracy}{varies on a per-country basis}.
}
\details{
City-level geolocation
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{geo_country}} for country-level identification, \code{\link{geo_tz}}
for tzdata-compatible timezone identification and \code{\link{geo_netspeed}} for connection
type detection.
}

