% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geo_country}
\alias{geo_country}
\title{geo_country}
\usage{
geo_country(ip_addresses)
}
\arguments{
\item{ips}{a vector of IP addresses}
}
\value{
a vector of country names. NULL or invalid responses from the API will be replaced with the string "Invalid".
}
\description{
Country-level geolocation
}
\details{
\code{geo_country} geolocates IP addresses to the country level, providing the
\href{https://en.wikipedia.org/wiki/ISO_3166-2}{ISO 3166-2} code for the resulting country.
It uses \href{http://dev.maxmind.com/geoip/}{MaxMind's binary geolocation database}.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{geo_city}} for city-level identification, \code{\link{geo_region}} for region-level,
\code{\link{geo_tz}} for tzdata-compatible timezone identification and \code{\link{geo_netspeed}} for connection
type detection.
}

