% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geo_netspeed}
\alias{geo_netspeed}
\title{Connection type detection}
\usage{
geo_netspeed(ips)
}
\arguments{
\item{ips}{a vector of IP addresses}
}
\value{
a vector of connection categories. NULL or invalid responses from the API will be replaced with
the string "Invalid".
}
\description{
\code{geo_netspeed} provides connection types for the provided IP address(es)
It uses (in order) \href{http://dev.maxmind.com/geoip/}{MaxMind's binary geolocation database}
and \href{https://github.com/maxmind/GeoIP2-python}{the associated Python API}, with
\code{\link{rpy}} as a connector. The Python API is required for it to work.

Unlike the other geo-related functions, \code{geo_netspeed} does not currently support IPv6
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{geo_city}} for city-level identification, \code{\link{geo_tz}}
for tzdata-compatible timezone identification, or \code{\link{geo_country}} for country-level geolocation.
}

