% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geo_region}
\alias{geo_region}
\title{geo_region}
\usage{
geo_region(ip_addresses)
}
\arguments{
\item{ips}{a vector of IP addresses. These will be processed through \code{xff_handler}
before being run, so don't worry if they're a bit groaty.}
}
\value{
a vector of region names. NULL or invalid responses from the API will be replaced with the string "Invalid".
}
\description{
\code{geo_region} geolocates IPv4 and IPv6 addresses to provide region-level results, providing
the English-language name for that region (or a digit-based name in the case that administrative
districts in the pertinent nation do not have proper nouns). It uses
\href{http://dev.maxmind.com/geoip/}{MaxMind's binary geolocation database} - the only
limitation is that accuracy
\href{http://www.maxmind.com/en/city_accuracy}{varies on a per-country basis}.
}
\details{
Region-level geolocation
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{geo_country}} for country-level identification, \code{\link{geo_city}} for city-level
geolocation, \code{\link{geo_tz}} for tzdata-compatible timezone identification and \code{\link{geo_netspeed}}
for connection type detection.
}

