% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geo_tz}
\alias{geo_tz}
\title{geo_tz}
\usage{
geo_tz(ip_addresses)
}
\arguments{
\item{ips}{a vector of IP addresses. These will be processed through \code{xff_handler}
before being run, so don't worry if they're a bit groaty.}
}
\value{
a vector of region names. NULL or invalid responses from the API will be replaced with the string "Invalid".
}
\description{
\code{geo_tz} returns the timezone, in a tzdata-compliant format, associated with an IPv4 IP address.
This can then be used to (for example) localise server-side timestamps. It uses
\href{http://dev.maxmind.com/geoip/}{MaxMind's binary geolocation database} - the only
limitation is that accuracy
\href{http://www.maxmind.com/en/city_accuracy}{varies on a per-country basis}.
}
\details{
tzdata-compatible timezone retrieval
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{geo_country}} for country-level identification, \code{\link{geo_city}} for city-level
geolocation, \code{\link{geo_region}} for region-compatible timezone identification and \code{\link{geo_netspeed}}
for connection type detection.
}

