% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hive_range}
\alias{hive_range}
\title{hive_range}
\usage{
hive_range(start_ts, end_ts)
}
\arguments{
\item{start_ts}{the start timestamp, as a POSIXlt or POSIXct object}

\item{end_ts}{the end timestamp, as a POSIXlt or POSIXct object}
}
\value{
a character string capable of being pushed into a query with \code{\link{paste}} that contains
"WHERE year=...".
}
\description{
year/month/day partition generation for hive
}
\details{
\code{hive_range} generates an approximation of the smallest range of year/month/day partitions
needed to scan to get all of the requests between a starting and ending timestamp.

The key word is "approximate"; it's not as efficient as writing these clauses by hand. But that's not always an
option, particularly with automated software.
}
\seealso{
\code{\link{hive_query}} for querying hive, and \code{\link{to_log}} and \code{\link{log_strptime}} for
handling hive date/time objects.
}

