% Generated by roxygen2 (4.0.2): do not edit by hand
\name{log_sieve}
\alias{log_sieve}
\title{log_sieve}
\usage{
log_sieve(log_data)
}
\arguments{
\item{log_data}{an input data.frame or data.table. This should ideally be the output of \code{\link{sampled_logs}}
or \code{\link{hive_query}}, since some of \code{log_sieve}'s arguments are name-specific rather than
indices-specific.}
}
\value{
a data.table containing those rows of \code{log_data} that are pageviews.
}
\description{
Prototype pageviews filter for the Wikimedia request logs
}
\details{
\code{log_sieve} contains the prototype filter for "pageviews", as applicable to the Wikimedia request logs.
It consumes logs, tags the "actual" pageviews, and returns them. While it's there, the XFFs are also
passed through to the ip_address field, replacing those IPs. It's implemented in R, so the full definition
can be seen just by printing \code{log_sieve}.

\code{log_data}, the first argument
}
\seealso{
code{\link{sampled_logs}}, to read from the sampled logs, or \code{\link{hive_query}} to read from
the HDFS-based logs.
}

