% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multi_with_tz}
\alias{multi_with_tz}
\title{multi_with_tz}
\usage{
multi_with_tz(dt, timestamp_col, timezone_col, handle_timestamps = FALSE,
  ts_handler)
}
\arguments{
\item{dt}{the data.table of timestamps and timezones (having other columns is not a problem)}

\item{timestamp_col}{the name of the column containing timestamps}

\item{timezone_col}{the name of the column containing tzdata-compatible timezones.}

\item{handle_timestamps}{whether or not timestamps need to be converted into a POSIX format from something
else. Set to FALSE by default}

\item{ts_handler}{if handle_timestamps is TRUE, a function used for the conversion. To handle MediaWiki
timestamps, for example, we'd use ts_handler = mw_strptime.}
}
\description{
localise timestamps to a range of different timezones
}
\details{
\code{multi_with_tz} takes a data.table of timezones and timestamps, and localises the timestamps
to match the newly-specified timezones.
}
\seealso{
\code{\link{geo_tz}} for extracting tzdata-compatible timezones from IP addresses,
\code{\link{log_strptime}} and \code{\link{mw_strptime}} for request log/MediaWiki timestamp
handling.
}

