% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mw_strptime}
\alias{mw_strptime}
\title{timestamp handling for MediaWiki}
\usage{
mw_strptime(x)
}
\arguments{
\item{x}{a vector containing full (14-character) MediaWiki timestamps}
}
\value{
a vector of POSIXlt timestamps
}
\description{
Format timestamps from our MediaWiki dbs as POSIXlt timestamps.
}
\seealso{
\code{\link{mysql_query}} for querying a single database, \code{\link{global_query}}
for querying multiple databases, and code{\link{log_strptime}} for equivalent functionality
as applied to the RequestLogs.
}

