\name{mw_to_posix}
\alias{mw_to_posix}
\title{
MediaWiki/POSIX timestamp conversion
}
\description{
Converts full MediaWiki timestamps to POSIXlt timestamps
}
\usage{
mw_to_posix(timestamps)
}
\arguments{
  \item{timestamps}{
A vector of 14-character MediaWiki timestamps. Can be numeric or character strings in nature.}
}

\seealso{
\code{\link{posix_to_mw}} for converting POSIX objects to MediaWiki timestamps
}
