% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mysql_delete}
\alias{mysql_delete}
\title{Wrapper for removing rows from a MySQL table, on the matching of some conditional}
\usage{
mysql_delete(db, table, conditional = NULL)
}
\arguments{
\item{db}{a database}

\item{table}{the name of the table}

\item{conditional}{a conditional statement (beginning 'WHERE...') for the row deletion. NULL by
default, which deletes the entire contents of the table.}
}
\value{
A TRUE (or an error)
}
\description{
\code{mysql_delete} is a simple wrapper around RMySQL that allows a useR to remove records from a table in
one of the analytics-store.eqiad.wmnet databases.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{mysql_query}} for reading from the same range of databases, \code{\link{mysql_write}}
for writing to them, or \code{\link{mysql_exists}} for checking the existence of a table.
}

