% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mysql_exists}
\alias{mysql_exists}
\title{Wrapper for checking the existence of a MySQL table}
\usage{
mysql_exists(db, table_name)
}
\arguments{
\item{db}{a database}

\item{table_name}{the name of the table}
}
\value{
A TRUE or FALSE confirming whether the table exists
}
\description{
\code{mysql_write} is a simple wrapper around RMySQL that allows a useR to check table existence in a database on
analytics-store.eqiad.wmnet.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{mysql_query}} for reading from the same range of databases, or \code{\link{mysql_write}}
for writing to them.
}

