% Generated by roxygen2 (4.0.2): do not edit by hand
\name{namespace_match_generator}
\alias{namespace_match_generator}
\title{namespace name/number retrieval}
\usage{
namespace_match_generator(to_return = FALSE)
}
\arguments{
\item{to_return}{whether or not to return the dataset, as well as save it to file.}
}
\value{
Either the value TRUE, if to_return is FALSE, or a data.table containing the project dbname,
language name, project type, namespace ID, and localised namespace name.
}
\description{
\code{namespace_match_generator} generates a localised list of namespace names and the equivalent IDs,
and saves it to file. This is then used by \code{\link{namespace_match}}
}
\details{
namespace_match_generator queries the various Wikimedia APIs to build up a dataset of what the localised
namespace names are on each project. For this reason it needs internet access - don't bother using it on
stat1002, for example.
}
\seealso{
\code{\link{namespace_match}}, which uses the resulting dataset to perform namespace name/ID matching.
}

