% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_uuids}
\alias{parse_uuids}
\title{parse_uuids}
\usage{
parse_uuids(url_strings)
}
\arguments{
\item{url_strings}{a vector of URLs from the request logs.}
}
\value{
a vector of unique IDs, or the string "Invalid" where no valid ID could be identified.
}
\description{
parse App UUIDs out of requestlog URLs
}
\details{
\code{parse_uuids} takes URLs from the request logs and parses through them hunting for
the unique identifiers included in API requests by the Wikimedia Apps. These are then extracted and provided
as a vector.

In the event that a valid ID cannot be identified, the string "Invalid" is instead returned.
}
\seealso{
\code{\link{sampled_logs}} and \code{\link{hive_query}} for retrieving request logs;
\code{\link{log_strptime}} for handling request log date/times.
}

