% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paste0}
\alias{paste0}
\title{paste without a separator}
\usage{
paste0(...)
}
\arguments{
\item{...}{arguments to pass into paste()}
}
\description{
What it says on the tin. Necessary for jsonlite and does not exist in all
R versions, sadly
}

