% Generated by roxygen2 (4.0.2): do not edit by hand
\name{referer_handler}
\alias{referer_handler}
\title{referer_handler}
\usage{
referer_handler(urls)
}
\arguments{
\item{urls}{a vector of URLs}
}
\value{
a vector of hostnames or organisational handles.
}
\description{
extract the host from a URL
}
\details{
\code{referer_handler} extracts the host from a URL and extracts general organisational
handles ("google", "reddit")
}

