\name{rpy_vec}
\alias{rpy_vec}
\title{
rpy results to vector conversion
}
\description{
Convert the results of rpy calls into a vector}
\usage{
rpy_vec(list)
}
\arguments{
  \item{list}{
  The list you want converted into a vector
}
}

\seealso{
\code{\link{rpy_df}} for conversion of rpy results into vectors, and \code{\link{rpy}} itself.
}