% Generated by roxygen2 (4.0.2): do not edit by hand
\name{session_length}
\alias{session_length}
\title{session_length}
\usage{
session_length(intertimes, local_minimum = 3600L)
}
\arguments{
\item{local_minimum}{the threshold (in seconds) to split out a new session on. Set to 3600
by default.}

\item{timestamps}{a vector of inter-time values.}
}
\value{
a vector of session length counts, in seconds.
}
\description{
Session counting function
}
\details{
\code{session_length} takes a vector of intertime values (generated via \code{\link{intertimes}},
or in any other way you see fit), splits them into sessions, and calculates the approximate
length (in seconds) of each session. It's implemented in C++ since, even with the hinkiness
of my C++, it's still several OoM faster than the equivalent R (recursion in R is dumb).
}
\section{Warning}{

Unlike \code{\link{intertime}} or \code{\link{session_count}}, the methodology here has not been
settled on; the internal workings of this function can and probably will change.
}
\seealso{
\code{\link{intertimes}}, for generating inter-time values, and \code{\link{session_count}} for
simply counting the number of sessions.
}

