% Generated by roxygen2 (4.0.2): do not edit by hand
\name{session_pages}
\alias{session_pages}
\title{session_pages}
\usage{
session_pages(intertimes, local_minimum = 3600L)
}
\arguments{
\item{intertimes}{a vector of intertime values.}

\item{local_minimum}{the threshold at which to split a series of intertime values into
multiple sessions. Set to 3600 seconds (1 hour) by default.}
}
\description{
count the number of pages in a session (or multiple sessions)
}
\details{
\code{session_pages} counts the number of pages in a session, or in multiple
sessions, based on a provided numeric vector of intertime values, which can be
generated via \code{\link{intertimes}}.
}
\seealso{
\code{\link{intertimes}} for generating intertime values, \code{\link{session_length}}
for session length, and \code{\link{session_count}} for the number of distinct sessions represented
by a series of intertime values.
}

