% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_log}
\alias{to_log}
\title{to_log}
\usage{
to_log(timestamp)
}
\arguments{
\item{timestamp}{a vector of POSIX timestamps, represented as an object of class POSIXct, POSIXlt or
character.}
}
\value{
a vector of log-compatible timestamps, stored as character strings.
}
\description{
POSIX-to-log timestamp conversion
}
\details{
\code{to_mw} takes a POSIXct or POSIXlt timestamp (or a character representation of those formats)
and converts it into a character string that matches our RequestLogs' timestamp format, as used
in the sampled and unsampled logs.
}
\seealso{
\code{\link{hive_query}} and \code{\link{sampled_logs}} for accessing the RequestLogs,
\code{\link{log_strptime}} for the same operation in reverse, and \code{\link{to_mw}} for
the equivalent operation for MediaWiki date/times.
}

