% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_mw}
\alias{to_mw}
\title{to_mw}
\usage{
to_mw(timestamp)
}
\arguments{
\item{timestamp}{a vector of POSIX timestamps, represented as an object of class POSIXct, POSIXlt or
character.}
}
\value{
a vector of MediaWiki-compatible timestamps, stored as character strings
}
\description{
POSIX-to-MediaWiki timestamp conversion
}
\details{
\code{to_mw} takes a POSIXct or POSIXlt timestamp (or a character representation of those formats)
and converts it into a character string consumable by MediaWiki - and, by extension, consumable
by our MySQL databases that contain MediaWiki-derived content.
}
\seealso{
\code{\link{mysql_query}} and \code{\link{global_query}} for querying our MySQL databases,
\code{\link{mw_strptime}} for the same operation in reverse, and \code{\link{to_hive}} for
the equivalent operation for our Hive datastore date/times.
}

