% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wiki_crawler}
\alias{wiki_crawler}
\title{wiki_crawler}
\usage{
wiki_crawler(agents)
}
\arguments{
\item{agents}{a vector of user agents}
}
\value{
a logical vector indicating whether or not a user agent was identified as a MediaWiki-specific
spider.
}
\description{
identifies Wikimedia-specific crawlers from their user agents
}
\details{
\code{\link{ua_parse}} is great for identifying spiders, but only /generic/ spiders - for
obvious reasons. Unfortunately there are some Wikimedia-specific spiders out there which need to be
caught. \code{wiki_crawler} hopes to identify these.
}
\seealso{
\code{\link{hive_query}} and \code{\link{sampled_logs}} for extracting request logs,
\code{\link{ua_parse}} for generic user-agent identification.
}

