\name{WMUtils-package}
\alias{WMUtils-package}
\alias{WMUtils}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab WMUtils\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-03\cr
License: \tab What license is it under?\cr
}
}
\author{
Who wrote it


}