% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geo_city}
\alias{geo_city}
\title{City-level geolocation}
\usage{
geo_city(ips)
}
\arguments{
\item{ips}{a vector of IP addresses}
}
\description{
\code{geo_city} geolocates IPv4 and IPv6 addresses to provide city-level results, providing
the English-language name for that city. It uses (in order)
\href{http://dev.maxmind.com/geoip/}{MaxMind's binary geolocation database} and
\href{https://github.com/maxmind/GeoIP2-python}{the associated Python API}, with
\code{\link{rpy}} as a connector. The Python API is required for it to work - the only other
limitation is that accuracy
\href{http://www.maxmind.com/en/city_accuracy}{varies on a per-country basis}.

NULL or non-country responses from the API will be replaced with the string "Invalid".
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{geo_country}} for country-level identification and \code{\link{geo_tz}}
for tzdata-compatible timezone identification.
}

