% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geo_tz}
\alias{geo_tz}
\title{tzdata-compatible timezone retrieval}
\usage{
geo_tz(ips)
}
\arguments{
\item{ips}{a vector of IP addresses}
}
\description{
\code{geo_tz} returns the timezone, in a tzdata-compliant format, associated with an IPv4 IP address.
This can then be used to (for example) localise server-side timestamps.
It uses (in order) \href{http://dev.maxmind.com/geoip/}{MaxMind's binary geolocation database}
and \href{https://github.com/maxmind/GeoIP2-python}{the associated Python API}, with \code{\link{rpy}} as a connector.
The Python API is required for it to work - the only other limitation is that accuracy
\href{http://www.maxmind.com/en/city_accuracy}{varies on a per-country basis.}
NULL or non-country responses from the API will be replaced with the string "Invalid".
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{geo_city}} for city-level identification and \code{\link{geo_country}}
for country-level identification
}

