% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hive_query}
\alias{hive_query}
\title{Run a query against the WMF hive instance}
\usage{
hive_query(query, ...)
}
\arguments{
\item{query}{a query, or the location of a .hql file containing a query.}

\item{...}{other arguments to pass to read.delim.}
}
\value{
a data.frame containing the results of the query.
}
\description{
\code{hive_query} is a simple wrapper around the command line that makes queries
against our Hive/Hadoop infrastructure more convenient.
}

