% Generated by roxygen2 (4.0.2): do not edit by hand
\name{host_handler}
\alias{host_handler}
\title{host_handler}
\usage{
host_handler(urls, strict = TRUE)
}
\arguments{
\item{urls}{a vector of URLs}

\item{strict}{whether or not to extract org handles; set to TRUE by default. This is mostly useful for referer
parsing.}
}
\value{
a vector of hostnames or organisational handles.
}
\description{
extract the host from a URL
}
\details{
\code{host_handler} extracts the host from a URL and, if asked, also strictly filters the resulting hosts,
extracting general organisational handles (for example "Google", or "Reddit").
}

