% Generated by roxygen2 (4.0.2): do not edit by hand
\name{host_handler}
\alias{host_handler}
\title{host_handler}
\usage{
host_handler(urls)
}
\arguments{
\item{urls}{a character vector of URLs}
}
\value{
a vector of hostnames, or "Unknown" if the hostname was invalid
}
\description{
\code{host_handler} is a domain-neutral host extractor for URLs, excluding generic prefixes (http, https, www)
and paths/queries.
}
\details{
extracts the hostname, TLD and subdomains from a generic URL
}
\seealso{
\code{\link{project_extractor}} for extracting Wikimedia language codes and projects.
}

