% Generated by roxygen2 (4.0.2): do not edit by hand
\name{intertimes}
\alias{intertimes}
\title{intertimes}
\usage{
intertimes(timestamps)
}
\arguments{
\item{timestamps}{a vector of timestamps. These can be POSIXlt, POSIXct, numeric/integer (if you've
already converted them into seconds), or either MediaWiki or RequestLog timestamps.}
}
\description{
calculate inter-time periods between timestamps
}
\details{
\code{\link{intertimes}} takes a set of timestamps and generates the interval between them, in seconds,
having sorted the timestamps from earliest to latest.
}
\seealso{
\code{\link{session_count}}, which takes a set of intertime periods and work out how many sessions
they represent, \code{\link{session_length}}, which works out the approximate length
(in seconds) of each session, or \code{\link{session_pages}}, which works out how many pages
are represented (split into sessions) by a series of intertime values.
}

