% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mysql_query}
\alias{mysql_query}
\title{Wrapper for querying the R&D MySQL instances}
\usage{
mysql_query(query, db, dt = TRUE)
}
\arguments{
\item{query}{the SQL query you want to run.}

\item{db}{the name of the database you want to run the query against.}

\item{dt}{whether to return the results as a data.table. TRUE by default.}
}
\value{
A data.frame or data.table containing the results of the query.
}
\description{
\code{mysql_query} is a simple wrapper around RMySQL that allows a useR to query the dbs on
analytics-store.eqiad.wmnet.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{global_query}} for querying multiple "production" databases in
analytics-store.eqiad.wmnet with a common query, \code{\link{mw_strptime}} for converting
MediaWiki timestamps into POSIXlt timestamps.
}

