\name{rpy_vec}
\alias{rpy_vec}
\title{rpy-to-vector}
\usage{
rpy_vec(list)
}
\arguments{
  \item{list}{the rpy-produced list you want converted into
  a vector #'}
}
\description{
\code{rpy_vec} converts the results of rpy calls into a
vector after sanitising them for missing values
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}
\seealso{
\code{\link{rpy_df}} for conversion of rpy results into
vectors, and \code{\link{rpy}} itself.
}

