% Generated by roxygen2 (4.0.2): do not edit by hand
\name{session_length}
\alias{session_length}
\title{session_length}
\usage{
session_length(x, local_minimum = 3600)
}
\arguments{
\item{local_minimum}{the threshold (in seconds) to split out a new session on. Set to 3600
by default.}

\item{timestamps}{a vector of inter-time values.}
}
\value{
a vector of session length counts, in seconds.
}
\description{
Session counting function
}
\details{
\code{session_length} takes a vector of intertime values (generated via \code{\link{intertimes}},
or in any other way you see fit), splits them into sessions, and calculates the approximate
length (in seconds) of each session.
}
\seealso{
\code{\link{intertimes}}, for generating inter-time values, and \code{\link{session_count}} for
simply counting the number of sessions.
}

