% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ua_parse}
\alias{ua_parse}
\title{ua_parse}
\usage{
ua_parse(user_agents)
}
\arguments{
\item{user_agents}{A vector of unparsed user agents}
}
\value{
a data.frame containing the results of the UA parsing - the device, the operating system,
the browser, and the browser's major and minor version numbers.
}
\description{
user agent parsing
}
\details{
\code{ua_parse} acts as a connector to the C++ implementation of
\href{https://github.com/ua-parser/uap-cpp}{tobie's ua-parser}, consuming a vector of user agents
and returning a data frame, with each field (see 'arguments') as a distinct column.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}

