% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ua_parse}
\alias{ua_parse}
\title{User agent parsing}
\usage{
ua_parse(user_agents, fields = c("device", "os", "browser", "browser_major",
  "browser_minor"))
}
\arguments{
\item{user_agents}{A vector of unparsed user agents}

\item{fields}{The elements you'd like to return. Options are "device" (the device code, when known),
"os" (the operating system, when known), "browser" (the browser), "browser_major" (the major version of the browser)
and browser_minor (the minor version of the browser)}
}
\description{
\code{ua_parse} acts as a connector to
\href{https://github.com/tobie/ua-parser}{tobie's ua-parser}, consuming a vector of user agents
and returning a data frame, with each field (see 'arguments') as a distinct column.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}

