% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xff_handler}
\alias{xff_handler}
\title{xff_handler}
\usage{
xff_handler(ips)
}
\arguments{
\item{ips}{a vector of IP chains}
}
\value{
a vector of either the last IP in a chain, or the IP if no chain was present
}
\description{
extracts the last IP address from a chain of IPs reported through the x_forwarded_for field
}

